requirements.txt
crispy-bootstrap5==0.6
Django==4.0.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
